requirements.txt
absl-py==0.11.0
appdirs==1.4.4
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.2.0
autopep8==1.5.7
backcall==0.2.0
black==21.7b0
bleach==3.2.1
cachetools==4.2.1
certifi==2021.5.30
cffi==1.14.3
click==8.0.1
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
flake8==3.8.4
google-auth==1.27.1
google-auth-oauthlib==0.4.3
grpcio==1.36.1
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.2
Jinja2==2.11.2
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.0
nbformat==5.0.7
nest-asyncio==1.4.1
notebook==6.1.4
numpy==1.21.0
oauthlib==3.1.0
opencv-python==4.5.3.56
packaging==20.4
pandocfilters==1.4.2
parso==0.7.1
pathspec==0.8.1
pickleshare==0.7.5
Pillow==8.3.1
prometheus-client==0.8.0
prompt-toolkit==3.0.7
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pywinpty==0.5.7
pyzmq==19.0.2
regex==2020.11.13
requests-oauthlib==1.3.0
rope==0.18.0
rsa==4.7.2
scipy==1.7.0
Send2Trash==1.5.0
six==1.16.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.2.0
toml==0.10.1
tomli==1.0.4
torch-tb-profiler==0.2.1
tqdm==4.61.2
traitlets==5.0.4
wcwidth==0.2.5
webencodings==0.5.1
wincertstore==0.2
wrapt==1.11.2
yapf==0.30.0
